requirements.txt
environs==11.0.0
Django==5.0.4
pytest-django==4.8.0
pydantic==1.10.0
